implement status change logic in StatusBtn component + minor change in CurrentCarModal
