+++ src/components/CurrentCarsItem/CurrentCarsItem.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/display-name */
-import { useState, useEffect, useMemo } from "react";
+import { useEffect, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import toast from "react-hot-toast";
 import clsx from "clsx";
@@ -11,24 +11,19 @@ import { BsUiChecksGrid } from "react-icons/bs";
 
 import { getCurrentCars } from "../../redux/cars/operations.js";
 import { selectCurrentCars } from "../../redux/cars/selectors.js";
+import { selectSelectedServiceId } from "../../redux/auth/selectors.js";
 import { calculateTimeInService } from "../../utils/calculateTimeInService.js";
 import { getStatusDetails } from "../../utils/getStatusDetails.js";
 
 import absentAutoImg from "../../assets/images/absentAutoImg.webp";
-import CurrentCarModal from "../Modals/CurrentCarModal/CurrentCarModal";
-import StatusBtn from "../../shared/StatusBtn/StatusBtn.jsx";
-import Modal from "../Modals/Modal/Modal.jsx";
+import StatusBtn from "../sharedComponents/StatusBtn/StatusBtn.jsx";
 import styles from "./CurrentCarsItem.module.css";
-import { selectSelectedServiceId } from "../../redux/auth/selectors.js";
 
 export default function CurrentCarsItem() {
   const dispatch = useDispatch();
   const currentCars = useSelector(selectCurrentCars);
 
-  const selectedServiceId = useSelector(selectSelectedServiceId); // необхідно для коректної роботи вибору сервісів
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [selectedCar, setSelectedCar] = useState(null);
+  const selectedServiceId = useSelector(selectSelectedServiceId);
 
   useEffect(() => {
     if (!selectedServiceId) {
@@ -44,18 +39,8 @@ export default function CurrentCarsItem() {
           error.message
         );
       });
-  }, [dispatch, selectedServiceId]); // необхідно для коректної роботи вибору сервісів
+  }, [dispatch, selectedServiceId]);
 
-  const closeModal = () => {
-    setIsModalOpen(false);
-    setSelectedCar(null);
-  };
-
-  const handleStatusChange = () => {
-    dispatch(getCurrentCars());
-  };
-
-  // Мемоізація іконки в залежності від статусу
   const getStatusIcon = useMemo(() => {
     return (status) => {
       switch (status) {
@@ -95,7 +80,6 @@ export default function CurrentCarsItem() {
     };
   }, []);
 
-  // Мемоізація списку автомобілів для уникнення зайвих обчислень
   const renderedCars = useMemo(() => {
     return currentCars?.map((car) => {
       const icon = getStatusIcon(car.status);
@@ -124,11 +108,7 @@ export default function CurrentCarsItem() {
             </h4>
           </div>
           <div className={styles.detailsContainer}>
-            <StatusBtn
-              car={car}
-              setIsModalOpen={setIsModalOpen}
-              setSelectedCar={setSelectedCar}
-            />
+            <StatusBtn car={car} />
             {/* тут буде імпорт кнопки деталей від Валі */}
             <div className={styles.statusContainer}>
               <p className={clsx(styles.carStatus, className)}>
@@ -141,19 +121,5 @@ export default function CurrentCarsItem() {
     });
   }, [currentCars, getStatusIcon]);
 
-  return (
-    <>
-      {renderedCars}
-      {isModalOpen && (
-        <Modal isOpen={isModalOpen} onClose={closeModal}>
-          <CurrentCarModal
-            car={selectedCar}
-            status={selectedCar?.status}
-            onClose={closeModal}
-            onStatusChange={handleStatusChange}
-          />
-        </Modal>
-      )}
-    </>
-  );
+  return <>{renderedCars}</>;
 }
