+++ src/components/sharedComponents/DetailsBtn/DetailsBtn.module.css
